[native] Create a ChangeRolesHeaderLeftButton to cancel role changes
Summary:
Per the designs, we want the button on the top left of the screen to be 'Cancel'. This diff introduces the `headerLeft` component.

I followed some of the work for the invite links modal on `native` to get the types correct for creating a separate component for `headerLeft` and `headerRight` (later in the stack), so also adding @tomek as a non-blocking reviewer for these 3 diffs

Depends on D8157

Test Plan:
Confirm that the 'Cancel' button appears on the top left of the header and closes the screen

{F580217}

Reviewers: atul, ginsu, tomek

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8158