[native] Create a ChangeRolesHeaderLeftButton to cancel role changes
Summary: Per the designs, we want the button on the top left of the screen to be 'Cancel'. This diff introduces the `headerLeft` component. I followed some of the work for the invite links modal on `native`
to get the types correct

Depends on D8157

Test Plan: Confirm that the 'Cancel' button appears on the top left of the header and closes the screen

Reviewers: atul, ginsu

Subscribers: