[native] Create a ChangeRolesHeaderRightButton to save role changes
Summary: This diff creates the 'save' headerRight component that is used to actually process the role change. When save is clicked, one of two things happen:

(1) If a member's role is changed, change the role and close the screen

(2) If a member's role remains unchanged, just close the screen

Because the save button is located in the react navigation header component, the best way I found to handle keeping track of the data from the role change screen was to access the info from `route.params`
after the role change screen sets them via `navigation.setParams`. If there is a better way to do this though, I'm open to changing this approach. I used [[
https://github.com/react-navigation/react-navigation/issues/145 | this github issue ]] as a reference for this approach.

Depends on D8159

Test Plan: Please see the video below showing changing a member's role

Reviewers: atul, ginsu, tomek

Subscribers: