[native] Introduce RegistrationTerms screen
Summary:
This diff begins the stack to addresses [ENG-3951](https://linear.app/comm/issue/ENG-3951/add-terms-of-use-privacy-policy-acceptance-to-new-registration-flow).

[This screen](https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?type=design&node-id=13233%3A268284&t=cbRKwoO1TefPCOhM-1) will be used for making sure the user is accepting our terms of service when creating an account.

This diff just sets up the scaffolding for the page. Later diffs will add functionality.

I opted not to include any usage of this screen in this diff, as it would've added too much complexity to this initial diff to handle moving the registration code. That will be handled in the next diff.

Depends on D8154

Test Plan: Here's what it looks like now: [screenshot]

Reviewers: atul, ginsu

Subscribers: