[landing] fix investor card description container height
Summary:
I actually did find a small bug with the investor card height. The height was still 56px, but when we decreased the font size/line height this 56px was too big. This diff fixes this bug by making the value of the height dynamically change as well based on the screen width. To determine these clamp values the same strategy from D8173 was used here.

Depends on D8173

Test Plan:
Please see the before and after

Before:
{F581456}

After:
{F581457}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8174