[native] Add error handling to KeyserverSelection
Summary:
I want to display an error to the user if we aren't able to reach the keyserver that they've selected.

Depends on D8197

Test Plan:
First, I overrode `defaultURLPrefix` in `native/utils/url-utils.js` to return some random string. That made sure that the "default" of Ashoat's keyserver does not work, and makes it so the registration flow only worked if the user selected a custom keyserver and all of the server calls in the registration flow used that custom keyserver.

Next, I played around with the screen: {F585701}

Reviewers: atul, ginsu

Reviewed By: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8198