[web] Change alert type in UserListItem
Summary:
A new version of: https://phab.comm.dev/D5583 (to avoid rebasing problems).

Raised in D5518, it doesn't make sense for only one of alertText and alertTitle to be defined. This diff changes the type to make sure they are always defined together and updates the usages.

Test Plan: Checked for flow errors. Check if thread composer works correctly on web and native.

Reviewers: inka, michal

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8223