[keyserver] Allow routing to invite endpoint
Summary: When landing page is configured with `/` base path, it receives all the requests before `/invite/:secret` can be handled - this route should be declared before any other router.

Test Plan: Configure landing page so that it is server on `/`. Open `/invite/abc` route and check if a user is redirected to `{commBasePath}/handle/invite/abc`. This testing procedure still is far from our prod setup, but should be enough to check if the solution is correct.

Reviewers: bartek, kamil, inka, ted