[native] introduce emoji keyboard component
Summary: This should be a simple move diff. Created an emoji keyboard component to abstract out all the additional functionality we will add to the emoji keyboard like search, adding the themes, etc.

Test Plan: Verified and confirmed that the emoji keyboard still opens up and works in the with the different message tooltips and emoji avatar creation

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8210