[native] Introduce `NativeEditThreadAvatarProvider`
Summary:
Slightly different than what we did with `NativeEditUserAvatarProvider`. In that case we just needed to pass in platform-specific functions which we could do from `root.js`.

However, in `EditThreadAvatarProvider` we use `NavContext` in order to get the `activeThreadID`. Introduced `NativeEditThreadAvatarProvider` to get `activeThreadID` and pass it to `EditThreadAvatarProvider`.

In next diff will rename `EditThreadAvatarProvider` to `BaseEditThreadAvatarProvider` and will move to `lib`.

---

Depends on D8219

Test Plan: Things are still working as expected w/ setting thread avatars.

Reviewers: ashoat, ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8220