[web] Introduce placeholder `EditUserAvatar` component
Summary:
Introduce placeholder `EditUserAvatar` component and include in `AccountSettings` if `editingAvatarsOnWebEnabled` is `true`.

This diff is super minor and `EditUserAvatar` is effectively an alias for `UserAvatar`.

I'm building this component incrementally in this stack and I think it made sense to separate this from subsequent diffs despite how trivial this one is.

---

Depends on D8221

Test Plan:
Made sure that `EditUserAvatar` component appeared as expected after manually setting `editingAvatarsOnWebEnabled` to `true`.

Reviewers:
ashoat, ginsu