[web] Display invite links button conditionally
Summary: When a community has invite link, we display it next to the community in the drawer. We're using our Menu component in a new variant.

Depends on D8236

Test Plan: Check if button is displayed only for communities with invite links and if clicking it results in menu being shown.

Reviewers: bartek, kamil, inka