add search bar to emoji keyboard
Summary:
This diff adds the search bar to the emoji keyboard. To do this, I read through the docs and found this [[ https://thewidlarzgroup.github.io/rn-emoji-keyboard/docs/documentation/Examples/search | example here ]].

One thing to note is that searching for emoji sometimes is a little strange (take a look at the demo below), after App.js I did get on a call with the owners of this library, and did share with them that this is something this should be improved. They said they would work on it, so hoping this will be improved soon "for free". If we have other points of feedback too, I do have an open line of communication with their team and they are very friendly/open to feedback.
{F587463}

https://linear.app/comm/issue/ENG-3791/mobile-emoji-keyboard-search

Test Plan:
Please view the demo video below

{F587431}

Reviewers: atul, kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8211