[web] Add message search modal
Summary:
issue: https://linear.app/comm/issue/ENG-3462/add-modal-for-searching-messages
For now the server is called after every key press, this will be changed in the future. Some elements like the "search" button and the footers are not yet implemented. For now the query is not stored
anywhere

{F591283}

Test Plan:
opened the modal and checked that it works - typed in query is sent to the server and the results show up. Checked that when the user scrolls to the bottom new messages are fetched.

I checked that all tooltip actions work

Reviewers: kamil, kuba, michal

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8240