[landing] Monitor WalletConnect modal height
Summary:
See parent diff:

> As part of the WalletConnect 2 update, RainbowKit is now using WalletConnect's [Web3Modal](https://web3modal.com/).
>
> This modal changes sizes more frequently, and as a result later diffs will complicate the code for monitoring the WalletConnect modal.

This diff implements the aforementioned "complicated" logic for monitoring the height of the WalletConnect modal.

Depends on D8251

Test Plan:
Tested with the rest of the stack. I tested all three SIWE endpoints:

1. Web
2. Native (old flow)
3. Native (new flow)

For each, I tested both Rainbow and WalletConnect (both of which need WalletConnect), and navigated around the WalletConnect modal to various places.

Reviewers: atul, ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8252