Manually serialize notification data to avoid passing encrypted notification to JavaScript
Summary: The context for this issue is here: https://linear.app/comm/issue/ENG-4146/navigation-to-thread-after-tapping-notification-is-broken-on-android. RemoteMessage instance caches its serialized form. Therefore even if we change in-memory data and display
correct unencrypted notification, when we pass RemoteMessage instance to JS via Intent we use cached serialized encrypted notification. Therefore we have to manually serialize unencrypted notification before sending it to JS.

Test Plan: Before applying this differential after tapping notification user will not be navigated to the correct thread. After applying this differential this will be fixed.

Reviewers: tomek, bartek

Subscribers: