Revert encrypted Android notification to avoid thread navigation regression
Summary:
This differential prevents keyserver from sending encrypted notification to Android devices. The reason is temporary regresion that user is not directed to thread after tapping notification. The regression is already handled by suitable differential
but unit it is not landed we have to stop encryption notifications.

Test Plan: Build the Android app. Send notification, tap it. Ensure that you are directed to ther particular thread.

Reviewers: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8247