[landing] Fix detection of RainbowKit modal dismissal
Summary:
After the RainbowKit update, this was broken and was detecting modal dismissal when the WalletConnect modal was opened.

To address this, I added some state that tracks whether the WalletConnect modal is open, and disabled the dismiss behavior while it's open.

Depends on D8252

Test Plan:
Tested with the rest of the stack. I tested all three SIWE endpoints:

1. Web
2. Native (old flow)
3. Native (new flow)

For each, I tested both Rainbow and WalletConnect (both of which need WalletConnect), and navigated around the WalletConnect modal to various places.

Reviewers: atul, ginsu

Subscribers: