[native] Update native to match new landing SIWE behavior
Summary:
I separated this out into its own commit to make it easier to test the old `native` with the new `landing`. I wanted to make sure we didn't break old clients.

1. Resize `BottomSheet` to specific values when the WalletConnect modal is open.
2. Style the handle to match the `WebView` when the WalletConnect modal is open.

Depends on D8253

Test Plan:
Tested with the rest of the stack. I tested all three SIWE endpoints:

1. Web
2. Native (old flow)
3. Native (new flow)

For each, I tested both Rainbow and WalletConnect (both of which need WalletConnect), and navigated around the WalletConnect modal to various places.

I also tested both the new `native` client (as of this diff) as well as the old one (as of the previous diff). The new client resizes better, but the old client is still usable.

Reviewers: atul, ginsu