[web] Intirduce onClearText to Search
Summary:
issue: https://linear.app/comm/issue/ENG-3466/add-some-state-for-keeping-the-search-query
In message search when a user clears the query we will need to remove the query from context. For this I need to ba able to pass a callback to the `Search` component. The `flex: 1` was added because in
message search modal we want to have a button next to it, and have the two of them take up the full width of the modal.

Test Plan:
Tested that a provided callback is called on clear

Reviewers:
kamil, kuba, michal
Subscribers: