[web] Prevent nested scroll in the message results modal
Summary: We want to prevent nested scrolls inside of the message results modal. This means that for the modal, we want to display the messages in full length (and keep the modal scrollable). However, in the
toggle pin modal (the modal that shows the message and allows the user to pin/unpin a message), we want to keep that scrollable since otherwise a long message could overflow off the screen and make the action
buttons unclickable.

The easiest solution I thought of for this was to maintain a `scrollable` prop, set to `true` if we want to allow the message result to be scrollable if it is long, and `false` if we want to display the
message in full.

Test Plan: Confirm that the toggle pin modal allows the user to scroll through the message, but the message results modal shows it in full.

Reviewers: michal, inka

Subscribers: