[native] Prevent nested scroll in the message results screen
Summary:
This is the native counterpart to the previous diff that handled the nested scroll in the message results modal on web. These changes will maintain the scrollable message in the toggle pin modal, but show the message in full when navigating to see the pinned messages or search results.

The changes here also cover the issue where two pinned messages back to back would prevent the user from scrolling to see other messages [[
https://linear.app/comm/issue/ENG-4150/scrolling-issue-in-pinned-messages-message-search#comment-c020d958  | ENG-4150 ]].

Depends on D8259

Test Plan:
Checked the toggle pin modal to confirm messages still scroll, checked both the search results / pinned messages screens to confirm that the messages show in full, and confirm that two long messages won't block scroll now.

{F597666}

Reviewers: michal, inka

Reviewed By: inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8262