[web] Implement a custom dropdown menu component
Summary:
The discussion was held on Linear (https://linear.app/comm/issue/ENG-4038/create-a-modal-that-allows-admins-to-change-a-members-role), but we needed a dropdown menu to match the Figma designs for
displaying potential roles that can be assigned to a member in the modal that will come down the stack. This diff creates a `Dropdown` component that is general enough to be reused for different use cases
down the line, if needed.

Depends on D8135

Test Plan: The actual dropdown menu will be verified in a later diff

Reviewers: atul, ginsu

Reviewed By: atul, ginsu

Subscribers: ashoat, tomek, ted

Differential Revision: https://phab.comm.dev/D8136