[web] Add emoji `Picker` to `EmojiAvatarSelectionModal`
Summary:
Add emojimart `Picker` to `EmojiAvatarSelectionModal` and hook up to `pendingAvatarEmoji`.

NOTE: Styling here isn't final. There weren't detailed designs for this screen so I'm going to improvise a bit. That said, going to focus on functionality first so disregard how it looks for now.

---

Depends on D8267

Test Plan:
Works as expected:

{F598068}

Reviewers: ashoat, ginsu, rohan, kamil, inka, michal

Reviewed By: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8268