[web] Wrap modals with `WebEditUserAvatarProvider`
Summary:
We want to be able to access eg `setUserAvatar` from `EmojiAvatarSelectionModal`, however, we can't access the context since the modal is being rendered outside of "mainContent".

We're only wrapping modals with `WebEditUserAvatarProvider` when the user is logged in.

---

Depends on D8268

Test Plan:
Able to access `EditUserAvatarContext` from `EmojiAvatarSelectionModal` as expected.

Reviewers:
ashoat, ginsu