[web] Prevent changing the only admin role in a community
Summary:
We need to prevent the only admin from removing themselves as `Admins` in a community. Based on the designs, we should display an error in the modal indicating why the saved role changes failed.

https://linear.app/comm/issue/ENG-4041/prevent-changing-the-only-admin-role-in-a-community

Depends on D8200

Test Plan:
Please see the video below covering the cases of changing the only admin role

{F575275}

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ted, ashoat, tomek

Differential Revision: https://phab.comm.dev/D8141