[web] Remove `LoadingIndicator` from `EditUserAvatar`
Summary:
In D8285 we introduced a `showSpinner` prop to the `Avatar` component. In this diff we remove the `LoadingIndicator` from `EditUserAvatar` component and instead pass `showSpinner` down to `Avatar` to take advantage of that.

This deduplicates redundant spinner code. Now that we have a "toggleable" spinner in `Avatar`, components derived from it can use that spinner "for free."

---

Depends on D8285

Test Plan:
Things continue to look as expected:

{F600929}

Reviewers: ashoat, ginsu, rohan

Reviewed By: rohan

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8286