[web] Remove `LoadingIndicator` from `EditUserAvatar`
Summary:
In D8285 we introduced a `showSpinner` prop to the `Avatar` component. In this diff we remove the `LoadingIndicator` from `EditUserAvatar` component and instead pass `showSpinner` down to `Avatar`.

This deduplicates redundant spinner code. Now that we have a "toggleable" spinner in `Avatar`, and components derived from it can use that spinner "for free."

---

Depends on D8285

Test Plan:
Things continue to look as expected:

Reviewers:
ashoat, ginsu, rohan