[web] Use local holder for blob deletion
Summary:
Replaces D8257 - on web, a newly-created blob holder needs to be kept to be able to delete the blob, when a user cancels sending a media while it is already uploaded

Depends on D8256

Test Plan: Enabled blob service uploads, selected a photo to send. Ensured the blob was uploaded successfully. Then clicked the cross button to unselect the photo. Ensured the blob was deleted

Reviewers: atul, ashoat

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8320