[keyserver] call identity service during account creation and deletion
Summary:
when we create a new account with a username, we should add it to the reserved usernames list
being maintained by the identity service.

when we delete an existing account with a username, we should remove it from the reserved usernames list being maintained by the identity service.

Depends on D8299

Test Plan:
tested this by creating and deleting a new user locally. saw that the username was added to and then removed from my ddb table in localstack.

also tried removing a reserved username that didn't exist in ddb and got back an error as expected.

Reviewers: atul!

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8300