[web] Add manage invite links menu item
Summary:
The new item is displayed conditionally based on the permissions and uses an SVG icon.

Depends on D8306

Test Plan:
Check if the new item is displayed for communities where user has permission.
Check if invite link button is displayed only when invite link exists.
Check if three dots button is displayed only when there's a menu item.

Reviewers: bartek, kamil, inka