[android] Bump `com.facebook.fresco:*` deps to `2.5.0`
Summary:
Basically it appears that `2.2.0`, which we previously specified in `build.gradle` was removed from all the artifact hosting services. It appears that `expo` depends on `fresco`, and the `fresco` version for Expo SDK 47 is `2.5.0` (see below)... so I'm bumping to that to see

1. If that resolves CI
2. If images continue to appear as expected on Android

{F602475}

Test Plan: CI, and trying out some flows on Android that involve images.

Reviewers: ashoat, ginsu, rohan, bartek

Reviewed By: bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8296