[lib/native] introduce getViewerAlreadySelectedMessageReactions
Summary:
This diff is the second part to highlight the already selected emojis to indicate on the emoji keyboard what emojis the user has already selected. This diff handles the case when `selectMultipleEmojis` is equal to `true`.

To handle this case we need to iterate through the `ReactionInfo` map and return an array of emojis that the user has already selected. I introduced a separate helper function called `getViewerAlreadySelectedMessageReactions` since `MultimediaMessageTooltipButton`, `TextMessageTooltipButton` and `RobotextMessageTooltipButton` will all need to do this similar functionality.

Here is a link to an example to the docs that helped me with this implementation: https://thewidlarzgroup.github.io/rn-emoji-keyboard/docs/documentation/Examples/selected_emojis

Linear Task: https://linear.app/comm/issue/ENG-3794/mobile-emoji-keyboard-highlight-emojis-that-a-user-has-already

Depends on D8284

Test Plan:
Please watch the demo video below:

{F601071}

Reviewers: atul, kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8288