[web] Introduce `large` variant of `ColorSelector`
Summary:
In this diff we introduce `size` prop to `ColorSelector` component and implement styling/layout for `large` variant.

This is for use in the `EmojiAvatarSelection` component.

Test Plan:
Here's how it looks:
https://blob.sh/f788e5.png

Also ensured that other usages of `ColorSelector` in `web` continue to appear as expected (eg ThreadSettings)

Reviewers:
ashoat, ginsu, rohan