[web] Fix message search results for old query appearing
Summary:
If the user searched for "first query", and then before the results had the chanse to appear, search for "second query", the results for "first query" would still be appended to the results array.
This diff fixes that by using an id for each server call, and after the results get fetched, the id of the server call is compared with the most recent id.

Test Plan:
Checked that flow is not broken in web and native that both use `useSearchMessages`. checked that the above scenario doen't happen anymore. Checked that it is still possible to fetch messages, and more
messages for the same query.

Reviewers:
kamil, kuba, michal
Subscribers: