[keyserver] Fetch primary links only when a user is logged in
Summary: Calling `userID` getter on `Viewer` results in an exception when a user isn't logged in. A logged out user shouldn't have access to invite links, so we can return an empty array for them.

https://linear.app/comm/issue/ENG-4185/cant-open-web-app-if-not-already-logged-in-dev#comment-6dba4247

Test Plan: Open web app as a logged out user and check if it loads correctly. Log in and check if the links are fetched.

Reviewers: kamil, inka, michal