[native] Show Roles option in community action sheet
Summary:
In the community action sheet, we want to introduce a 'Roles' option that will navigate to a screen listing all of the community roles and respective member counts. Since that involves navigation,
the `action` key/value will come later, and this just shows the option in the action sheet.

Depends on D8332

[[ https://linear.app/comm/issue/ENG-4166/show-roles-option-in-community-action-sheet | ENG-4166 ]]

Test Plan: Confirm that a 'Roles' option appears for those who have permissions to change roles, and it should not appear if otherwise.

Reviewers: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8333