[native] Create a roles navigator to handle the role creation flow
Summary:
Since the role creation flow will involve more than one screen, I believe it is necessary (or good to have since it's extendable) to have a `role-navigator` that will handle its related routes. This
diff creates the navigator, and the next few diffs in the stack will use this for navigation. I mainly used the invite links navigator & the community creation navigator as inspiration here.

Depends on D8333

[[ https://linear.app/comm/issue/ENG-4167/create-a-roles-navigator-to-handle-the-role-creation-flow | ENG-4167 ]]

Test Plan: The navigator was tested in the next few diffs as screens for the role creation flow are built out, but locally it works fine and are no issues.

Reviewers: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8335