[native] Create the community roles screen
Summary:
This diff handles creating the screen that will display all roles in a given community, alongside a count of how many members are assigned each role. There is also a button that will allow admins to create a new role - the functionality of the button will be implemented in the next diff. The designs for this screen can be found on the [[
https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?type=design&node-id=13217-292819&mode=design&t=dICx7mwEN7GpXXmf-0 | Figma ]].

Note: For now, I pull the `threadInfo` directly from the navigation params, but in a later diff I plan to pull the updated `threadInfo` from the redux store (a solution discussed in my 1:1 with Ashoat) in order to make sure the roles list is always kept up to date as new roles are added and users are navigated back to this screen. This will need to be done after the screen for creating a custom role is complete, so that's why it's going to be implemented later.

Depends on D8358

[[ https://linear.app/comm/issue/ENG-4168/create-the-community-roles-screen | ENG-4168 ]].

Test Plan:
Confirm that for communities, the correct roles and role counts show up in the new screen. Also verified that for communities with several custom roles, the scroll container looks fine.

{F610727}

Reviewers: atul, ginsu

Reviewed By: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8359