[native] Create the community roles screen
Summary:
This diff handles creating the screen that will display all roles in a given community, alongside a count of how many members are assigned each role. There is also a button that will allow admins to
create a new role - the functionality of the button will be implemented in the next diff. The designs for this screen can be found on the [[
https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?type=design&node-id=13217-292819&mode=design&t=dICx7mwEN7GpXXmf-0 | Figma ]].

Depends on D8358

[[ https://linear.app/comm/issue/ENG-4168/create-the-community-roles-screen | ENG-4168 ]].

Test Plan: Confirm that for communities, the correct roles and role counts show up in the new screen. Also verified that for communities with several custom roles, the scroll container looks fine.

Reviewers: atul, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8359