[native] Navigate to a screen to create a new role
Summary: Following the community roles screen, clicking the button should navigate to a new screen that will allow admins to create new roles with a set of permissions. The actual screen will be filled out in
a following diff, this just handles the navigation to what will be the create roles screen.

Test Plan: Confirmed that the navigation works and the animation from the previous screen to the new screen is good

Reviewers: atul, ginsu

Subscribers: