[native] Navigate to a screen to create a new role
Summary:
Following the community roles screen, clicking the button should navigate to a new screen that will allow admins to create new roles with a set of permissions. The actual screen will be filled out in a following diff, this just handles the navigation to what will be the create roles screen.

Depends on D8359

[[ https://linear.app/comm/issue/ENG-4170/navigate-to-a-screen-to-create-a-role | ENG-4170 ]]

Test Plan:
Confirmed that the navigation works and the animation from the previous screen to the new screen is good

{F612763}

Reviewers: atul, ginsu

Reviewed By: ginsu

Subscribers: ted, ashoat, tomek

Differential Revision: https://phab.comm.dev/D8379